extract withDefaultLayout to its own file
and use it in _app
